+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "tuck" extension will be documented in this file.
 
+## Unreleased
+
+- Use an explicit Python by fetching the workspace's Python where possible, this
+  aims to support Windows and or other environments where the workspace python
+  is not on $PATH
+
 ## 0.1.1
 
 - Update tuck library to 0.1.1: